+++ tests/sql_test_files/error_dynamic_component_expected_properties_of_type_array_or_object.sql
@@ -1,0 +1,1 @@
+select 'dynamic' as component; -- missing "properties" column