+++ tests/sql_test_files/error_invalid_json_in_dynamic_component.sql
@@ -1,0 +1,2 @@
+select 'dynamic' as component,
+    '{ "this object": "has a forbidden comma" , }' as properties; -- this is invalid JSON, and should cause an error